+++ dart/geobase/CHANGELOG.md
@@ -4,6 +4,8 @@ NOTE: Version 1.3.0 currently under development (1.3.0-dev.0).
 
 [geobase release 1.3.0](https://github.com/navibyte/geospatial/milestone/5)
 
+* [Add visual center point calculations for polygons #237](https://github.com/navibyte/geospatial/issues/237)
+
 🛠 Maintenance:
 * Add dependency on the `collections` package (for `PriorityQueue`).